backend/auth_service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==1.10.12
email-validator==2.0.0
motor==3.3.1
pymongo==4.5.0
bcrypt==4.0.1